nua: refactored the client side.
Now each handle has a queue of client-side requests; if a request is
pending, a new request from API is added to the queue. After the request is
complete, it is removed from the queue and destroyed by the default. The
exception is the client requests bound to a dialog usage: they are saved and
re-used when the dialog usage is refreshed (and sometimes when they are
terminated).

The client request is subclassed and its behaviour modified using virtual
function table in nua_client_methods_t.

darcs-hash:20070102121444-65a35-f1fa4d836b95611cfbc6141ff491afcb8b789a74.gz