nta.c: return 481 response to CANCEL if INVITE transaction was successful
According to the RFC 3261 state diagram the INVITE transaction is destroyed
immediately when a final 2XX series response is sent. Now nta returns a 481
response to CANCEL if it finds an INVITE server transaction that was
responded with a 2XX success response.

darcs-hash:20070214202456-55b16-26e78d85b04fa47ecf612966bd8766188390035a.gz