nua_session.c: delay transition to ready when O/A is incomplete
Delay sending ACK and subsequent transition of call to the ready state when
the 200 OK response to the INVITE is received if the SDP Offer/Answer
exchange using UPDATE/PRACK was still incomplete.

Previously, if the O/A using UPDATE or PRACK was incomplete and an 200 OK
was received, the call setup logic regarded this as a fatal error and
terminated the call.

Thanks for Mike Jerris for detecting and reporting this bug.

darcs-hash:20070723152956-65a35-36d194e38ec4ded0f20342489d55f2ab5594e2a5.gz