nua: added NUTAG_INITIAL_ROUTE() and NUTAG_INITIAL_ROUTE_STR()
Added tags NUTAG_INITIAL_ROUTE() and NUTAG_INITIAL_ROUTE_STR() that can be
used to determine the route set included in the request initiating dialogs
or outside them. They can be used instead or in addition to the
NUTAG_PROXY() when determining the outbound proxy.

The URL in NUTAG_PROXY() is not added to the requests, and it affects all
the requests, including in-dialog requests. The route set in
NUTAG_INITIAL_ROUTE{_STR}() is explictly added to the requests and it is
visible in the requests, and it affects only routing of the initial
requests.

TODO: interaction with service route set returned by registrar in the
Service-Route header.

darcs-hash:20070917141542-65a35-b7e7a38d4fbcc86ed156bfe5e353d91b37a8588d.gz